updated structure of project
