+++ app/structures/HashTable/node.py
